fix(setup): use a pointer-sized buffer for SendMessageTimeoutW's out-param
lpdwResult is PDWORD_PTR (DWORD_PTR = ULONG_PTR = pointer-width), which
is 8 bytes on 64-bit Windows. ctypes.c_ulong is always 4 bytes on
Windows (LLP64), so the OS call was writing 8 bytes into a 4-byte
buffer — an out-of-bounds write. ctypes.c_size_t matches DWORD_PTR's
width on both 32- and 64-bit Windows.

Found in adversarial review.

+++ claude/scripts/setup.py
@@ -762,7 +762,7 @@ def _set_os_env_var(key: str, value: str) -> bool:
             HWND_BROADCAST = 0xFFFF
             WM_SETTINGCHANGE = 0x001A
             SMTO_ABORTIFHUNG = 0x0002
-            result = ctypes.c_ulong()
+            result = ctypes.c_size_t()
             ctypes.windll.user32.SendMessageTimeoutW(
                 HWND_BROADCAST, WM_SETTINGCHANGE, 0, "Environment",
                 SMTO_ABORTIFHUNG, 5000, ctypes.byref(result),
